[SPARK-40448][CONNECT] Spark Connect build as Driver Plugin with Shaded Dependencies
This is a draft change of the current state of the Spark Connect prototype implemented as a driver plugin to separate the classpaths and shaded dependent libraries. The benefit of a driver plugin is that the plugin has access to everything inside Spark but can itself be a leaf-build. This allows us to shade it's own minimal set of dependencies without brining new dependencies into Spark.

Spark Connect is implemented as a driver plugin for Apache Spark. This means that it can be optionally loaded by specifying the following option during startup.

```
--conf spark.plugins=org.apache.spark.sql.sparkconnect.service.SparkConnectPlugin
```

The way that the plugin is integrated into the build system does not require any additional changes and the properly configured Jar files are produced as part of the build.

Spark Connect relies both on GRPC and Guava. Since the version of Guava is incompatible to the version used in Spark, the plugin shades this dependency directly as part of the build. In the case of Maven this is very easy in the case of SBT it's a bit trickier. For SBT we hook into the `copyDeps` stage of the build and instead of copying the artifact dependency we copy the shaded dependency instead.

https://issues.apache.org/jira/browse/SPARK-39375

Experimental API for Spark Connect

This patch was added by adding two new test surface areas. On the driver plugin side this patch adds rudimentary test infrastructure that makes sure that Proto plans can be translated into Spark logical plans. Secondly, there are Python integration tests that test the DataFrame to Proto conversion and test the end-to-end execution of Spark Connect queries.

To avoid issues with other parts of the build, only tests that require Spark Connect will be started with the driver plugin.

**Note:**
* The pyspark tests are only support for regular CPython and disabled for pypy
* `mypy` was initially disabled for this patch to reduce friction. Will be enabled again.

* `mypy` is disabled on the Python prototype until follow-up patches fix the warnings.
* The unit-test coverage in Scala is not good. Follow up tests will increases coverage, but to keep this test reasonable I did not add more code.

This patch adds GRPC, Protobuf, and an updated Guava to the leaf-build of Spark Connect. These dependencies are not available to the top-level Spark build and are not available to other Spark projects. To avoid dependency issues, the Maven and SBT build will shad GRPC, Protobuf and Guava explicitly.

Due to the size of the PR, smaller clean-up changes will be submitted as follow up PRs.

Closes #37710 from grundprinzip/spark-connect-grpc-shaded.

Lead-authored-by: Martin Grund <martin.grund@databricks.com>
Co-authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>